+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Add `smartwatch visualize` option `--role_order` for showing different annotation roles on different channel stacks.
 * Experimental new logic in `smartwatch project` for keyframe propagation / interpolation.
 * Added `polygon_simplify_tolerance` to tracking.
+* Add `--resolution` parameter to the tracker.
 
 ### Fixed
 * Switched scriptconfig objects to use "data=kwargs" rather than "default=kwargs" to avoid a copy issue.